+++ kernel/src/SystemCalls/SystemCall.cpp
@@ -115,6 +115,10 @@ extern "C" uint64_t SystemCallHandler(uint64_t num, uint64_t arg1, uint64_t arg2
         return 0;
     case SC_GETDIRENTS:
         return (uint64_t)(current_thread->sys$getdirents((fd_t)arg1, (struct dirent*)arg2, (size_t)arg3));
+    case SC_CHDIR:
+        return (uint64_t)(current_thread->sys$chdir((const char*)arg1));
+    case SC_FCHDIR:
+        return (uint64_t)(current_thread->sys$fchdir((fd_t)arg1));
     default:
         dbgprintf("Unknown system call. number = %lu, arg1 = %lx, arg2 = %lx, arg3 = %lx.\n", num, arg1, arg2, arg3);
         return -1;