+++ kernel/src/fs/VFS.hpp
@@ -38,6 +38,11 @@ struct VFS_MountPoint {
     Inode* RootInode;
 };
 
+struct VFS_WorkingDirectory {
+    VFS_MountPoint* mountpoint;
+    Inode* inode;
+};
+
 class FileStream; // defined in FileStream.hpp
 
 class VFS final : public FileSystem {
@@ -55,20 +60,24 @@ public:
 
     bool DeleteInode(FilePrivilegeLevel current_privilege, const char* path, bool recursive = false) override;
 
-    FileStream* OpenStream(FilePrivilegeLevel current_privilege, const char* path, uint8_t modes);
+    FileStream* OpenStream(FilePrivilegeLevel current_privilege, const char* path, uint8_t modes, VFS_WorkingDirectory* working_directory = nullptr);
     bool CloseStream(FileStream* stream);
 
-    DirectoryStream* OpenDirectoryStream(FilePrivilegeLevel current_privilege, const char* path, uint8_t modes);
+    DirectoryStream* OpenDirectoryStream(FilePrivilegeLevel current_privilege, const char* path, uint8_t modes, VFS_WorkingDirectory* working_directory = nullptr);
     bool CloseDirectoryStream(DirectoryStream* stream);
 
-    bool IsValidPath(const char* path) const;
+    bool IsValidPath(const char* path, VFS_WorkingDirectory* working_directory = nullptr) const;
 
-    Inode* GetInode(const char* path, FileSystem** fs = nullptr) const;
+    Inode* GetInode(const char* path, VFS_WorkingDirectory* working_directory = nullptr, FileSystem** fs = nullptr, VFS_MountPoint** mountpoint = nullptr) const;
+
+    VFS_MountPoint* GetMountPoint(FileSystem* fs) const;
+
+    VFS_WorkingDirectory* GetRootWorkingDirectory() const;
 
 private:
-    VFS_MountPoint* GetMountPoint(const char* path, Inode** inode = nullptr) const;
+    VFS_MountPoint* GetMountPoint(const char* path, VFS_WorkingDirectory* working_directory = nullptr, Inode** inode = nullptr) const;
 
-    bool isMountpoint(const char* path, size_t len); // When false is returned, the caller **MUST** check for any errors.
+    bool isMountpoint(const char* path, size_t len, VFS_WorkingDirectory* working_directory = nullptr); // When false is returned, the caller **MUST** check for any errors.
 
 private:
     VFS_MountPoint* m_root;